remove sampler
